Attempt to more precisely remove URLs.
Note that URLs contained within markdown mark-up are already removed.

Also this permits <SUBREDDIT> and <USERNAME> to be added to the wordlist when
they are of the form /r/<SUBREDDIT> and /u/<USERNAME> respectively. This is
desired behavior.

Also part of this change is to move possessive form elimination into the new
tokenize function which removes the possessiveness completely.

Finally, this completely removes the previous "exclude_words" behavior which
was to ignore anything if they contained one of those items. With these
changes, most of those exclusions are no longer necessary.